add another test case, clean up and organize fixtures [load test: skip]
