+++ tests/integration/conftest.py
@@ -4,5 +4,7 @@
 pytest_plugins = [
     "tests.integration.fixtures.gcs",
     "tests.integration.fixtures.metrics",
+    "tests.integration.fixtures.app_client",
+    "tests.integration.fixtures.manifest",
     "tests.integration.api.v1.curated_recommendations.corpus_backends.fixtures",
 ]
